[HELM] Only deploy master branch and tagged builds to snapshot repository

+++ helm/upload.sh
@@ -15,13 +15,16 @@ if [ "${TRAVIS_PULL_REQUEST}" = "false" ] && [ "${TRAVIS_TAG}" != "" ]; then
   rm -rf ./helm-charts
 fi
 
-cd helm-charts-snapshot
-git add .
-git remote rm origin
-git remote add origin https://phertweck:$GITHUB_API_KEY@github.com/FraunhoferIOSB/helm-charts-snapshot
-git commit -m "Travis build ${TRAVIS_BUILD_NUMBER} pushed"
-git push origin master -fq
-cd ../
-rm -rf ./helm-charts
+# Only deploy master branch and tagged builds to snapshot repository
+if [ "${TRAVIS_PULL_REQUEST}" = "false" ] && ([ "${TRAVIS_BRANCH}" == "master" ] && [ "${TRAVIS_TAG}" != "" ]); then
+    cd helm-charts-snapshot
+    git add .
+    git remote rm origin
+    git remote add origin https://phertweck:$GITHUB_API_KEY@github.com/FraunhoferIOSB/helm-charts-snapshot
+    git commit -m "Travis build ${TRAVIS_BUILD_NUMBER} pushed"
+    git push origin master -fq
+    cd ../
+    rm -rf ./helm-charts
+fi
 
 echo "Helm chart build and pushed"
